chore: ini msw providers
