+++ apps/app/package.json
@@ -32,6 +32,7 @@
     "@lightdotso/data": "workspace:*",
     "@lightdotso/hooks": "workspace:*",
     "@lightdotso/kysely": "workspace:*",
+    "@lightdotso/msw": "workspace:*",
     "@lightdotso/params": "workspace:*",
     "@lightdotso/prisma": "workspace:*",
     "@lightdotso/query": "workspace:*",
