+++ packages/const/src/index.ts
@@ -13,6 +13,11 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+export {
+  BASE_API_AUTHENTICATED_URL,
+  BASE_API_URL,
+  BASE_LOCAL_ADMIN_URL,
+} from "./api_urls";
 export { CHAIN_IDS } from "./chain_ids";
 export { CHAINS, MAINNET_CHAINS, TESTNET_CHAINS } from "./chains";
 export { CONTRACT_ADDRESSES } from "./contract_addresses";
